Fixed buffers' size
The reason for this is that when you allocate memory using malloc, you are requesting a contiguous block of memory of a specific size in bytes. If you allocate memory for sizeof(struct mystruct*) * buffers_size, you will only allocate memory for a list of pointers to struct mystruct objects, not for the objects themselves.